+++ lib/user/user_data.dart
@@ -3,6 +3,7 @@ class UserData {
   String? pfpBase64;
   int level;
   int expPoints;
+  String? username;
 
   // constructor
   UserData({
@@ -10,5 +11,6 @@ class UserData {
     this.pfpBase64,
     this.level = 1,
     this.expPoints = 0,
-  });
+    String? username,
+  }) : username = username ?? uid; // Default username is the UID
 }